+++ src/Pyro4/socketutil.py
@@ -311,8 +311,16 @@ def createSocket(bind=None, connect=None, reuseaddr=False, keepalive=True, timeo
                     if sock in sw:
                         break  # yay, writable now, connect() completed
                     elif sock in se:
+                        try:
+                            sock.close()  # close the socket that refused to connect
+                        except socket.error:
+                            pass
                         raise socket.error("connect failed")
             else:
+                try:
+                    sock.close()  # close the socket that refused to connect
+                except socket.error:
+                    pass
                 raise
     if keepalive:
         setKeepalive(sock)